User Auth 1.0
Please do not push this to prod, I would appreciate input.
This is userauth using supabase (communal github login) and sveltekit auth helpers. From what I can tell users are able to log in and log out.
Need to find a better way to check if the user is logged in than using local storage. I assume there is a way to get session that I am missing.

Ping me in a discord message with what I should change and improve upon.