User Auth 1.1
Fixed the 300000000 file issue.
The login works and supabase says that the user is logged in, I used local storage to check if the user is logged in but there has to be a better way like using getSession to see if the user is logged in.

I used client side session sharing:
https://supabase.com/docs/guides/auth/auth-helpers/sveltekit#client-side